Tarea Producto

+++ Models/Producto.cs
@@ -4,7 +4,7 @@ namespace demomvc.Models
     public class Producto
     {
         public string Nombre { get; set; }
-        public decimal Precio { get; set; }
-        public int Cantidad {get; set;}
+        public double Precio { get; set; }
+        public double Cantidad {get; set;}
     }
 }
